[skip ci] feat: cleanup some
